search/backend/requirements.txt
flask
flask_cors
PyYAML
pandas
elasticsearch
PyGithub
gunicorn  # For production WSGI server